requirements.txt
pip>=21.2.4
tracardi-plugin-sdk
tracardi
tracardi_dot_notation
pydantic
python-fullcontact
wheel
twine